examples/tests/requirements.txt
requests
gitpython